# apps/rbac/backend/openapi.yaml
openapi: 3.1.0
info:
  description: V1 of RBAC service.
  version: 1.0.0
  title: RBAC API
servers:
  - url: https://baumac.dev/api/v1
    description: Base URL for the RBAC service
security:
  - bearerAuth: []
paths:
...
  /permissions:
    get:
      summary: Get a list of permissions.
      operationId: list-permissions
      parameters:
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          $ref: '#/components/responses/ListPermissionsResponse'
      tags:
        - Permissions
    post:
      summary: Create a new permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermissionRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatePermissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
      tags:
        - Permissions
  /permissions/{id}:
    get:
      summary: Get a permission by ID.
      operationId: get-permission
      parameters:
        - schema:
            type: string
            format: uuid
          name: id
          in: path
          required: true
          description: |
            The ID of the role to get.
      responses:
        '200':
          $ref: '#/components/responses/GetPermissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Permissions
    patch:
      summary: Patch updatable properties of a permission.
      operationId: patch-permission
      parameters:
        - schema:
            type: string
            format: uuid
            example: bcd7681a-45fb-4171-b4ac-436d9756d4fa
          name: id
          in: path
          required: true
          description: |
            The ID of the permission to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchPermissionRequest'
      responses:
        '201':
          $ref: '#/components/responses/PatchPermissionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
      tags:
        - Permissions
    delete:
      summary: Delete a permission by ID.
      operationId: delete-permission
      parameters:
        - schema:
            type: string
            format: uuid
            example: bcd7681a-45fb-4171-b4ac-436d9756d4fa
          name: id
          in: path
          required: true
          description: |
            The ID of the permission to delete.
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Permissions
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      description: Determines which page of a collection API to retrieve.
      required: false
      in: query
      style: deepObject
      schema:
        $ref: '#/components/schemas/PageParameters'
  schemas:
    Error:
      title: Error
      description: Error schema.
      type: object
      properties:
        statusCode:
          type: integer
          required: true
          enum: [ 400, 401, 403, 404, 409 ]
          example: 404
          description: The HTTP status code.
        message:
          type: string
          required: true
          description: Details about the error response.
          example: Not Found.
    PaginatedMeta:
      type: object
      title: PaginatedMeta
      description: Returns pagination information
      properties:
        page:
          type: object
          additionalProperties: false
          description: Contains pagination query parameters and the total number of objects returned.
          properties:
            number:
              type: integer
              required: true
              example: 2
              description: Specifies the number of objects returned per page.
            size:
              type: integer
              required: true
              example: 10
              maximum: 100
              description: Specifies the size of the page.
            total:
              type: integer
              required: true
              maximum: 1000
              example: 100
              description: Integer representation of the amount of items returned.
    PageParameters:
      title: 'PageParameters'
      type: object
      additionalProperties: false
      properties:
        size:
          type: integer
          description: The number of items included per page.
          example: 10
        number:
          type: integer
          description: The specific page number of the collection results.
          example: 2
    Permission:
      title: Permission
      description: Permission schema.
      type: object
      properties:
        id:
          type: string
          required: true
          format: uuid
          description: ID of the permission.
          readOnly: true
        name:
          type: string
          required: true
          description: Name of the permission.
        description:
          type: string
          required: true
          nullable: true
          description: Description of the permission.
        created_at:
          type: string
          required: true
          format: date-time
          example: '2022-11-04T20:10:06.927Z'
          description: >-
            An RFC-3339 timestamp representation of permission creation date.
        updated_at:
          type: string
          required: true
          format: date-time
          example: '2022-11-04T20:10:06.927Z'
          description: >-
            An RFC-3339 timestamp representation of permission last updated date.
    CreatePermissionRequest:
      title: 'CreatePermissionRequest'
      description: Request schema for creating a permission.
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          required: true
          description: Name of the permission.
        description:
          type: string
          required: true
          nullable: true
          description: Description of the permission.
    PatchPermissionRequest:
      title: 'PatchPermissionRequest'
      description: Request schema for updating a permission.
      type: object
      additionalProperties: false
      properties:
        description:
          type: string
          required: false
          nullable: true
          description: Description of the permission.
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Unauthenticated
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
...
    CreatePermissionResponse:
      description: Response format for creating a permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Permission'
    GetPermissionResponse:
      description: Response format for getting a permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Permission'
    PatchPermissionResponse:
      description: Response format for updating a permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Permission'
    ListPermissionsResponse:
      description: A paginated list of permissions.
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
              meta:
                $ref: '#/components/schemas/PaginatedMeta'